simulation/requirements.txt
absl-py==2.1.0
astunparse==1.6.3
cachetools==5.3.3
certifi @ file:///croot/certifi_1671487769961/work/certifi
charset-normalizer==3.3.2
cloudpickle==2.2.1
cycler==0.11.0
fonttools==4.38.0
gast==0.3.3
google-auth==2.30.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.62.2
h5py==2.10.0
idna==3.7
imageio==2.31.2
importlib-metadata==6.7.0
joblib==1.3.2
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
Markdown==3.4.4
MarkupSafe==2.1.5
matplotlib==3.5.3
networkx==2.6.3
numpy==1.18.5
oauthlib==3.2.2
opencv-python==4.10.0.82
opt-einsum==3.3.0
packaging==24.0
Pillow==9.5.0
progressbar2==4.2.0
protobuf==3.20.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-utils==3.5.2
PyWavelets==1.3.0
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.4.1
six==1.16.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.3.0
tensorflow-gpu==2.3.0
tensorlayer==2.2.5
termcolor==2.3.0
threadpoolctl==3.1.0
tifffile==2021.11.2
typing_extensions==4.7.1
urllib3==2.0.7
Werkzeug==2.2.3
wrapt==1.16.0
zipp==3.15.0
